requirements.txt
Django>=5.1.4
python-decouple
djangorestframework
Pillow
django-filter
mysqlclient
django-redis
celery==5.3
requests==2.31
flake8==6.1.0
beautifulsoup4
pytest
pytest-django